Finished program main features, starting EF code
Right now the file tagging system is functional, with some minor issues fixed/updated.

Some features have been rethinked (are they useful or not) like adding a file and folder without tagging them, which would not be actually any better than just tagging the file.

Removing of files is still an option through buttons AND a system clean up feature (for unnecessary files).
Fixing files (when you migrate the database file into another computer) is a feature that is kept.
Some intentions of adding import / export features, but maybe after EF or WPF upgrade.

From here onward an entity framework version of SQLite will be added and tested. The EF will not be incorporated into anything but the tests for show casing that it can be implemented and can work very well.